move input handlers to separate file
