+++ src/main.rs
@@ -2,19 +2,19 @@ mod desync;
 mod frames;
 mod log_plugin;
 mod physics;
+mod rollback;
 
 use bevy::prelude::*;
 use bevy::tasks::IoTaskPool;
-use bevy_ggrs::{GGRSPlugin, PlayerInputs, Rollback, RollbackIdProvider, Session};
+use bevy_ggrs::{GGRSPlugin, Rollback, RollbackIdProvider, Session};
 use bevy_rapier2d::prelude::*;
-use bytemuck::{Pod, Zeroable};
 use desync::*;
 use frames::*;
-use ggrs::{Frame, InputStatus, PlayerHandle, PlayerType, SessionBuilder};
+use ggrs::{PlayerHandle, PlayerType, SessionBuilder};
 use log_plugin::LogSettings;
 use matchbox_socket::WebRtcSocket;
 use physics::*;
-use rand::{thread_rng, Rng};
+use rollback::*;
 
 use bevy::log::*;
 use bevy_inspector_egui::WorldInspectorPlugin;
@@ -48,12 +48,6 @@ const DESYNC_MAX_FRAMES: usize = 30;
 const MATCHBOX_ADDR: &str = "wss://match.gschup.dev/bevy-ggrs-rapier-example?next=2";
 // TODO: Maybe update this room name (bevy-ggrs-rapier-example) so we don't test with each other :-)
 
-// These are just 16 bit for bit-packing alignment in the input struct
-const INPUT_UP: u16 = 0b0001;
-const INPUT_DOWN: u16 = 0b0010;
-const INPUT_LEFT: u16 = 0b0100;
-const INPUT_RIGHT: u16 = 0b1000;
-
 /// Local handles, this should just be 1 entry in this demo, but you may end up wanting to implement 2v2
 #[derive(Default, Resource)]
 pub struct LocalHandles {
@@ -62,27 +56,6 @@ pub struct LocalHandles {
 
 #[derive(Default, Resource)]
 pub struct WebRtcSocketWrapper(pub Option<WebRtcSocket>);
-
-/// Our primary data struct; what players send to one another
-#[repr(C)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Pod, Zeroable)]
-pub struct GGRSInput {
-    // The input from our player
-    pub input: u16,
-
-    // Desync detection
-    pub last_confirmed_hash: u16,
-    pub last_confirmed_frame: Frame,
-    // Ok, so I know what you're thinking:
-    // > "That's not input!"
-    // Well, you're right, and we're going to abuse the existing socket to also
-    // communicate about the last confirmed frame we saw and what was the hash
-    // of the physics state.  This allows us to detect desync.  This could also
-    // use a new socket, but who wants to hole punch twice?  I have been working
-    // on a GGRS branch (linked below) that introduces a new message type, but
-    // it is not ready.  However, input-packing works good enough for now.
-    // https://github.com/cscorley/ggrs/tree/arbitrary-messages-0.8
-}
 
 /// The main GGRS configuration type
 #[derive(Debug)]
@@ -121,22 +94,6 @@ impl DeterministicSpawnBundle {
 #[derive(Copy, Clone, PartialEq, Eq, Debug, Default, Component)]
 pub struct Player {
     pub handle: usize,
-}
-
-/// Easy way to detect rollbacks
-#[derive(Default, Reflect, Hash, Resource, PartialEq, Eq)]
-#[reflect(Hash, Resource, PartialEq)]
-pub struct LastFrameCount {
-    pub frame: Frame,
-}
-
-/// Controls whether our opponent will inject random inputs while inactive.
-/// This is useful for testing rollbacks locally and can be toggled off with `r`
-/// and `t`.
-#[derive(Default, Reflect, Hash, Resource, PartialEq, Eq)]
-#[reflect(Hash, Resource, PartialEq)]
-pub struct RandomInput {
-    pub on: bool,
 }
 
 /// Our GameState, which will be rolled back and we will use to restore our
@@ -588,155 +545,6 @@ pub fn startup(
     let task_pool = IoTaskPool::get();
     task_pool.spawn(message_loop).detach();
     commands.insert_resource(WebRtcSocketWrapper(Some(socket)));
-}
-
-pub fn input(
-    _handle: In<PlayerHandle>, // Required by bevy_ggrs
-    keyboard_input: Res<Input<KeyCode>>,
-    random: Res<RandomInput>,
-    physics_enabled: Res<PhysicsEnabled>,
-    mut hashes: ResMut<FrameHashes>,
-    validatable_frame: Res<ValidatableFrame>,
-) -> GGRSInput {
-    let mut input: u16 = 0;
-    let mut last_confirmed_frame = ggrs::NULL_FRAME;
-    let mut last_confirmed_hash = 0;
-
-    // Find a hash that we haven't sent yet.
-    // This probably seems like overkill but we have to track a bunch anyway, we
-    // might as well do our due diligence and inform our opponent of every hash
-    // we have This may mean we ship them out of order.  The important thing is
-    // we determine the desync *eventually* because that match is pretty much
-    // invalidated without a state synchronization mechanism (which GGRS/GGPO
-    // does not have out of the box.)
-    for frame_hash in hashes.0.iter_mut() {
-        // only send confirmed frames that have not yet been sent that are well past our max prediction window
-        if frame_hash.confirmed
-            && !frame_hash.sent
-            && validatable_frame.is_validatable(frame_hash.frame)
-        {
-            info!("Sending data {:?}", frame_hash);
-            last_confirmed_frame = frame_hash.frame;
-            last_confirmed_hash = frame_hash.rapier_checksum;
-            frame_hash.sent = true;
-        }
-    }
-
-    // Do not do anything until physics are live
-    if !physics_enabled.0 {
-        return GGRSInput {
-            input,
-            last_confirmed_frame,
-            last_confirmed_hash,
-        };
-    }
-
-    if keyboard_input.pressed(KeyCode::W) {
-        input |= INPUT_UP;
-    }
-    if keyboard_input.pressed(KeyCode::A) {
-        input |= INPUT_LEFT;
-    }
-    if keyboard_input.pressed(KeyCode::S) {
-        input |= INPUT_DOWN;
-    }
-    if keyboard_input.pressed(KeyCode::D) {
-        input |= INPUT_RIGHT;
-    }
-
-    if input == 0 && random.on {
-        let mut rng = thread_rng();
-        // Return a random input sometimes, or maybe nothing.
-        // Helps to trigger input-based rollbacks from the unplayed side
-        match rng.gen_range(0..6) {
-            0 => input = INPUT_UP,
-            1 => input = INPUT_LEFT,
-            2 => input = INPUT_DOWN,
-            3 => input = INPUT_RIGHT,
-            _ => (),
-        }
-    }
-
-    GGRSInput {
-        input,
-        last_confirmed_frame,
-        last_confirmed_hash,
-    }
-}
-
-pub fn apply_inputs(
-    mut query: Query<(&mut Velocity, &Player)>,
-    inputs: Res<PlayerInputs<GGRSConfig>>,
-    mut hashes: ResMut<RxFrameHashes>,
-    local_handles: Res<LocalHandles>,
-    physics_enabled: Res<PhysicsEnabled>,
-) {
-    for (mut v, p) in query.iter_mut() {
-        let (game_input, input_status) = inputs[p.handle];
-        // Check the desync for this player if they're not a local handle
-        // Did they send us some goodies?
-        if !local_handles.handles.contains(&p.handle) && game_input.last_confirmed_frame > 0 {
-            log::info!("Got frame data {:?}", game_input);
-            if let Some(frame_hash) = hashes
-                .0
-                .get_mut((game_input.last_confirmed_frame as usize) % DESYNC_MAX_FRAMES)
-            {
-                // Only update this local data if the frame is new-to-us.
-                // We don't want to overwrite any existing validated status
-                // unless the frame is replacing what is already in the buffer.
-                if frame_hash.frame != game_input.last_confirmed_frame {
-                    frame_hash.frame = game_input.last_confirmed_frame;
-                    frame_hash.rapier_checksum = game_input.last_confirmed_hash;
-                    frame_hash.validated = false;
-                }
-            }
-        }
-
-        // On to the boring stuff
-        let input = match input_status {
-            InputStatus::Confirmed => game_input.input,
-            InputStatus::Predicted => game_input.input,
-            InputStatus::Disconnected => 0, // disconnected players do nothing
-        };
-
-        if input > 0 {
-            // Useful for desync observing
-            log::info!("input {:?} from {}: {}", input_status, p.handle, input)
-        }
-
-        // Do not do anything until physics are live
-        if !physics_enabled.0 {
-            continue;
-        }
-
-        let horizontal = if input & INPUT_LEFT != 0 && input & INPUT_RIGHT == 0 {
-            -1.
-        } else if input & INPUT_LEFT == 0 && input & INPUT_RIGHT != 0 {
-            1.
-        } else {
-            0.
-        };
-
-        let vertical = if input & INPUT_DOWN != 0 && input & INPUT_UP == 0 {
-            -1.
-        } else if input & INPUT_DOWN == 0 && input & INPUT_UP != 0 {
-            1.
-        } else {
-            0.
-        };
-
-        if horizontal != 0. {
-            v.linvel.x += horizontal * 10.0;
-        } else {
-            v.linvel.x = 0.;
-        }
-
-        if vertical != 0. {
-            v.linvel.y += vertical * 10.0;
-        } else {
-            v.linvel.y = 0.;
-        }
-    }
 }
 
 pub fn update_matchbox_socket(commands: Commands, mut socket_res: ResMut<WebRtcSocketWrapper>) {
